refactor(server): move get my user dto
