+++ server/src/Application/UseCases/User/GetMyUser/GetMyUserDTO.ts
