+++ server/src/Graphql/Resolvers/UserResolver.ts
@@ -1,8 +1,9 @@
 import { Context } from "@Database/Context";
-import type { GraphQLContext } from "@Graphql/GraphQLContext";
 import { UserRepository } from "@Repositories/UserRepository";
+import type { GraphQLContext } from "@Graphql/GraphQLContext";
+
 import type { ListUsersDTO } from "@UseCases/User/DTOs/ListUsersDTO";
-import type { GetMyUserDTO } from "@UseCases/User/DTOs/GetMyUserDTO";
+import type { GetMyUserDTO } from "@UseCases/User/GetMyUser/GetMyUserDTO";
 import type { UpdateUserDTO } from "@UseCases/User/DTOs/UpdateUserDTO";
 
 import { ListUsersUseCase } from "@UseCases/User/ListUsers/ListUsersUseCase";
